% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipByStateId.R
\name{bipByStateId}
\alias{bipByStateId}
\title{Get BIP (Broadband Initiatives Program) by State Ids.}
\usage{
bipByStateId(stateId = NULL)
}
\arguments{
\item{stateId}{A character or vector of characters with FIPS State Numeric Code. http://www.census.gov/geo/reference/ansi_statetables.html}
}
\value{
A dataframe with following information: http://www.broadbandmap.gov/broadbandmap/api/dictionary/v1/bip?format=json.
}
\description{
This function allows the user to find the BIP funding allocated to states by specifying the state fips code.
}
\examples{
bipByStateId(c("12","18","21"))
}

