% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipByStateName.R
\name{bipByStateName}
\alias{bipByStateName}
\title{Get BIP (Broadband Initiatives Program) by State Names.}
\usage{
bipByStateName(stateName = NULL)
}
\arguments{
\item{stateName}{A character or vector of characters with names of States.}
}
\value{
If \code{stateName="nation"} returns national bip
}
\description{
Get BIP (Broadband Initiatives Program) by State Names.
}
\examples{
bipByStateName(stateName="florida")
bipByStateName(stateName=c("South Carolina", "florida"))
bipByStateName(stateName="nation")
}

