% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censusByCoord.R
\name{censusByCoord}
\alias{censusByCoord}
\title{Get census information from coordinates.}
\usage{
censusByCoord(lat = NULL, long = NULL, geographyType = NULL)
}
\arguments{
\item{lat}{A numeric value, latitude}

\item{long}{A numeric value, longitude}

\item{geographyType}{A character. Type of geography.}
}
\value{
specific information about census.
}
\description{
Get census information from coordinates.
}
\examples{
censusByCoord(lat=25.7766074, long = -80.3714869, geographyType="block")
}

