% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censusByFips.R
\name{censusByFips}
\alias{censusByFips}
\title{This function finds the geography of a specified geography type by geography id.}
\usage{
censusByFips(geography = "state", geographyId = "01")
}
\arguments{
\item{geography}{A character with geography type. It can be state, county, tract, block, congdistrict, statehouse, statesenate, censusplace, msa, tribal, usf.}

\item{geographyId}{A character or vector of character with the corresponding geographical ID.}
}
\value{
A dataframe with following information: http://www.broadbandmap.gov/broadbandmap/api/dictionary/v1/census?format=json
}
\description{
It finds the geography of a specified geography type by geography id within the entire United States.
}
\examples{
censusByFips(geography="state", geographyId="12")
}

