% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censusByGeo.R
\name{censusByGeo}
\alias{censusByGeo}
\title{This function finds all the geographies specified by a geography name.}
\usage{
censusByGeo(geography = NULL, name = NULL, maxresults = 5)
}
\arguments{
\item{geography}{A character with geography type. It can be state, county, tract, block, congdistrict, statehouse, statesenate, censusplace or msa.}

\item{name}{A character with at least 3 leading characters.}

\item{maxresults}{specify the maximum results to be returned - defaults to 100.}
}
\value{
A dataframe with following information: http://www.broadbandmap.gov/broadbandmap/api/dictionary/v1/census?format=json
}
\description{
This function finds all the geographies specified by a geography name (e.g., Washington) of a specific geography type (e.g., congressional district) within the entire United States.
}
\examples{
censusByGeo(geography="county", name="Miami", maxresults=5)
}

