% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographicsByGeo.R
\name{demographicsByGeo}
\alias{demographicsByGeo}
\title{This function allows the user to search for the demographic information by ID.}
\usage{
demographicsByGeo(version = "jun2014", geography = NULL,
  geographyId = NULL)
}
\arguments{
\item{version}{A character with data version. Examples: jun2014, dec2013, jun2013, dec2012, jun2012, dec2011.}

\item{geography}{A character with geography type. It can be censusplace, msa, usf, county, statesenate, statehouse, congdistrict.}

\item{geographyId}{A character or vector of characteres with geographicals ID.}
}
\value{
A dataframe with following information: http://www.broadbandmap.gov/broadbandmap/api/dictionary/v1/demographics?format=json
}
\description{
This function allows the user to search for the demographic information for a particular geography type and geography ID from http://www.broadbandmap.gov/.
}
\examples{
demographicsByGeo(version = "jun2014", geography = "county",
geographyId = c("01001", "01003", "01005", "01007", "01011", "010015", "01017", "01013", "01025", "01035", "01023")
}

