% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolookup.R
\name{geolookup}
\alias{geolookup}
\title{Get information about .}
\usage{
geolookup(geography = "state", maxresults = 100)
}
\arguments{
\item{geography}{A character with type pf Geography. \code{county, censusplace, msa, usf, statesenate, statehouse, congdistrict, tribalnation}}

\item{geography}{A character with type pf Geography.}
}
\value{
If \code{stateName="nation"} returns national bip
}
\description{
Get information about .
}
\examples{
geolookup(geography = "state")
}

