% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipByStateId.R
\name{bipByStateId}
\alias{bipByStateId}
\title{Get BIP (Broadband Initiatives Program) by State Ids.}
\usage{
bipByStateId(stateId = NULL)
}
\arguments{
\item{stateId}{A character or vector of characters.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Get BIP (Broadband Initiatives Program) by State Ids.
}
\examples{
bipByStateId(c("12","18","21"))
}

