% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censusByCoord.R
\name{censusByCoord}
\alias{censusByCoord}
\title{Get census information by coordinates.}
\usage{
censusByCoord(lat = NULL, lon = NULL, geographyType = NULL)
}
\arguments{
\item{lat}{A numeric value, latitude}

\item{lon}{A numeric value, longitude}

\item{geographyType}{A character. Valid geography types: state, county, tract, block, congdistrict, statehouse, statesenate, censusplace, msa.}
}
\value{
A dataframe with following information: http://www.broadbandmap.gov/broadbandmap/api/dictionary/v1/census?format=json
}
\description{
This function returns the US Census Block geography ID information given a passed Latitude and Longitude. Latest data version released jun2014. Previous data versions include jun2011, dec2011, jun2012, dec2012, jun2013, dec2013.
}
\examples{
censusByCoord(lat=25.7766074, lon = -80.3714869, geographyType="block")
}

