% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countyBroadband.R
\name{countyBroadband}
\alias{countyBroadband}
\title{Search by county name for list of counties with response having a flag for fixed broadband availability greater than 50 percent for each county.}
\usage{
countyBroadband(county = NULL, date = "jun2014")
}
\arguments{
\item{county}{A character with a minimum of 1 character.}

\item{date}{A character with data version. Examples: jun2014, dec2013, jun2013, dec2012, jun2012, dec2011}
}
\value{
stateName countyName availabilityGt50PercentFlag countyId stateId
}
\description{
Search by county name for list of counties with response having a flag for fixed broadband availability greater than 50 percent for each county.
}
\examples{
countyBroadband(county="Miami")
}

